Handled conversion of movie files, restructured renaming and saving, and fixed bugs.

+++ MediaDirectoryConverter.py
@@ -4,25 +4,32 @@ from datetime import datetime
 from pillow_heif import register_heif_opener
 from PIL import Image
 from colorama import init, Fore, Style
+from ScoutDirectory import Scout
+from MovieConverter import movie_converter
 
 init()
 
 success_count = 0
 warning_count = 0
 error_count = 0
+current_file_count = 0
 
 register_heif_opener()
 
 
-def convert_files(folder_path, sub_directories_allowed, debug_enabled):
-    global success_count, warning_count, error_count
+def convert_files(folder_path, sub_directories_allowed, debug_enabled, scout):
+    global success_count, warning_count, error_count, current_file_count
 
     for filename in os.listdir(folder_path):
         if os.path.isdir(os.path.join(folder_path, filename)) and sub_directories_allowed:
             convert_files(os.path.join(folder_path, filename),
-                          sub_directories_allowed, debug_enabled)
+                          sub_directories_allowed, debug_enabled, scout)
         if os.path.isfile(os.path.join(folder_path, filename)):
-            if filename.lower().endswith((".jpg", ".mp4", ".png")):
+            if filename.split(".")[-1] != "json":
+                current_file_count += 1
+            print(Fore.BLUE +
+                  f"Converting file {current_file_count}/{scout.number_of_files}")
+            if filename.lower().endswith(tuple(scout.file_types)):
                 json_path = os.path.join(
                     folder_path, filename + ".supplemental-metadata.json")
 
@@ -37,13 +44,19 @@ def convert_files(folder_path, sub_directories_allowed, debug_enabled):
                             Fore.RED + f"ERROR: Failed to delete metadata {os.path.basename(json_path)}: {e}" + Style.RESET_ALL)
                         error_count += 1
 
-            if filename.lower().endswith(".heic"):
+            if filename.lower().endswith(tuple(scout.noncompatible_file_types)):
                 base_name = os.path.splitext(filename)[0]
+                extension = os.path.splitext(filename)[-1]
                 heic_path = os.path.join(folder_path, filename)
                 json_path = os.path.join(
-                    folder_path, base_name + ".HEIC.supplemental-metadata.json")
+                    folder_path, base_name + extension + ".supplemental-metadata.json")
 
-                new_filename = base_name + ".jpg"
+                if filename.lower().endswith(tuple(scout.noncompatible_video_file_types)):
+                    new_filename = movie_converter(
+                        os.path.join(folder_path, filename), debug_enabled)
+                    print(new_filename)
+                else:
+                    new_filename = base_name + ".jpg"
 
                 if os.path.exists(json_path):
                     try:
@@ -55,8 +68,21 @@ def convert_files(folder_path, sub_directories_allowed, debug_enabled):
 
                         if timestamp:
                             dt = datetime.fromtimestamp(int(timestamp))
-                            new_filename = dt.strftime(
-                                "%Y%m%d_%H%M%S") + ".jpg"
+                            if new_filename.lower().endswith(".mp4"):
+                                new_name = dt.strftime(
+                                    "%Y%m%d_%H%M%S")
+                                new_full_path = os.path.join(
+                                    folder_path, new_name)
+                                os.rename(new_filename, new_full_path + ".mp4")
+
+                                # Since I want to retain the original .MOV file
+                                os.rename(os.path.join(
+                                    folder_path, filename), new_full_path + ".mov")
+
+                                new_filename = new_full_path + ".mp4"
+                            else:
+                                new_filename = dt.strftime(
+                                    "%Y%m%d_%H%M%S") + ".jpg"
                     except Exception as e:
                         debug_enabled and print(
                             Fore.YELLOW + f"WARNING: Could not read metadata for {filename}: {e}" + Style.RESET_ALL)
@@ -65,15 +91,22 @@ def convert_files(folder_path, sub_directories_allowed, debug_enabled):
                 jpg_path = os.path.join(folder_path, new_filename)
 
                 try:
-                    with Image.open(heic_path) as img:
-                        img.convert("RGB").save(jpg_path, "JPEG")
-                    debug_enabled and print(
-                        Fore.GREEN + f"SUCCESS: Converted {filename} → {new_filename}" + Style.RESET_ALL)
+                    if not new_filename.lower().endswith(".mp4"):
+                        with Image.open(heic_path) as img:
+                            img.convert("RGB").save(jpg_path, "JPEG")
+                            debug_enabled and print(
+                                Fore.GREEN + f"SUCCESS: Converted {filename} → {new_filename}" + Style.RESET_ALL)
 
-                    os.remove(heic_path)
-                    debug_enabled and print(
-                        f"Deleted {filename}" + Style.RESET_ALL)
-
+                        os.remove(heic_path)
+                        debug_enabled and print(
+                            f"Deleted {filename}" + Style.RESET_ALL)
+                    else:
+                        debug_enabled and print(
+                            Fore.GREEN + f"SUCCESS: Converted {filename} → {new_filename}" + Style.RESET_ALL)
+                        # I want to retain the .MOV files since the conversion is not perfect but does the trick at least
+                        #  os.remove(heic_path)
+                        debug_enabled and print(
+                            f"Retained {filename}" + Style.RESET_ALL)
                     if os.path.exists(json_path):
                         os.remove(json_path)
                         debug_enabled and print(
@@ -106,7 +139,11 @@ def main():
 
     debug_enabled = debug_enabled_user_input == "y"
 
-    convert_files(folder_path, sub_directories_allowed, debug_enabled)
+    scout = Scout(sub_directories_allowed)
+    scout.scout_directory(folder_path)
+    print(f"Number of files: {scout.number_of_files}")
+
+    convert_files(folder_path, sub_directories_allowed, debug_enabled, scout)
 
     print(Fore.GREEN + f"Successful conversions: {success_count}")
     print(Fore.YELLOW + f"Warnings during conversion: {warning_count}")
